mempool: move mempoolIDs to sep file
